--- TODO
@@ -1,4 +1,0 @@
-- Make getter to report if sync is enabled instead of manually checking
-- Run code formatter on *.vue code
-- Sector shows when encrypted
-- Dev mode color
