+++ src/components/Popup/MenuPage.vue
@@ -26,6 +26,7 @@
 </template>
 <script lang="ts">
 import Vue from 'vue'
+import { syncTimeWithGoogle } from "../../popup";
 
 import IconArrowLeft from '../../../svg/arrow-left.svg'
 import IconInfo from '../../../svg/info.svg'
@@ -101,46 +102,4 @@ export default Vue.extend({
         }
     }
 })
-
-function syncTimeWithGoogle() {
-  return new Promise(
-    (resolve: (value: string) => void, reject: (reason: Error) => void) => {
-      try {
-        // tslint:disable-next-line:ban-ts-ignore
-        // @ts-ignore
-        const xhr = new XMLHttpRequest({ mozAnon: true });
-        xhr.open('HEAD', 'https://www.google.com/generate_204');
-        const xhrAbort = setTimeout(() => {
-          xhr.abort();
-          return resolve('updateFailure');
-        }, 5000);
-        xhr.onreadystatechange = () => {
-          if (xhr.readyState === 4) {
-            clearTimeout(xhrAbort);
-            const date = xhr.getResponseHeader('date');
-            if (!date) {
-              return resolve('updateFailure');
-            }
-            const serverTime = new Date(date).getTime();
-            const clientTime = new Date().getTime();
-            const offset = Math.round((serverTime - clientTime) / 1000);
-
-            if (Math.abs(offset) <= 300) {
-              // within 5 minutes
-              localStorage.offset = Math.round(
-                (serverTime - clientTime) / 1000
-              );
-              return resolve('updateSuccess');
-            } else {
-              return resolve('clock_too_far_off');
-            }
-          }
-        };
-        xhr.send();
-      } catch (error) {
-        return reject(error);
-      }
-    }
-  );
-}
 </script>
