+++ src/store/Qr.ts
