+++ src/store/Style.ts
