--- src/ui/add-account.ts
@@ -1,24 +1,0 @@
-import { OTPEntry } from '../models/otp';
-
-export async function addAccount() {
-  const ui = {
-    data: {},
-    // methods: {
-    //   addAccountManually: () => {
-    //     const entries = _ui.instance.entries as OTPEntry[];
-    //     for (let i = 0; i < entries.length; i++) {
-    //       // we have encrypted entry
-    //       // the current passphrase is incorrect
-    //       // shouldn't add new account with
-    //       // the current passphrase
-    //       if (this.$store.getters["accounts/currentlyEncrypted"]) {
-    //   this.$store.commit("notification/alert", this.i18n.phrase_incorrect);
-    //   return;
-    // }
-    //     }
-
-    //     _ui.instance.newAccount.show = true;
-    //   },
-    // },
-  };
-}
